Moved/added files for future use and legacy support memes
I moved all things that aren't actually being used in the website as of now into the "extra folder". In the future I plan on combining the jungleWallpaper and waterTexture in someway so that when you scroll up you see the jungle and the pets associated with it, and when you scroll down, you see the water along with all of the aquatic life.

-Removed Herobrine (XD)